Add flag for suppressing incsearch for PeculiarN
**Why** is the change needed?

As a part of #6 the peculiar commands were reworked as command-line
functions instead of using input, in order to support incremental
search.

One side effect is that PeculiarN shows the whole search as matching.
Put it behind a flag, since it's unclear whether this is preferred over
the previous implementation.

**How** is the need addressed?

- Add a flag called `g:peculiar#suppress_highlight_n`
- Use the old implementation if the flag is set.

Closes #7